Mailbox
Added more fake users and created the mailbox window and ways of adding new emails as the game goes on. You can also select the text, which I need to implement elsewhere, if I can.

Need to connect it to the game so you can open and close your email while investigating. Also auto-open the first email.