Waveform, Friends, and Button
Waveforms can have specified seeds so it matches for The Song. Random users check for any other users that are friends with them so there are backlinks. And fixed the soundwav button to only send you back to the main search hub when you click it when already in that screen (not mailbox or report).